# Swagger/doc/swagger-users.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Documentation",
    "description": "Documentation for users microservice"
  },
  "servers": [
    {
      "url": "http://localhost:3002/"
    }
  ],
  "tags": [
    {
      "name": "Users",
      "description": "Routes related to users"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get All Users.",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/getById": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get user by id",
        "parameters": [
          {
            "name": "userId",
            "description": "User Id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Register an User",
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users"
              }
            }
          }
        }
      }
    },
    "/update": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Update user.",
        "parameters": [
          {
            "name": "userId",
            "description": "User Id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users"
              }
            }
          }
        }
      }
    },
    "/delete": {
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "Delete user.",
        "parameters": [
          {
            "name": "userId",
            "description": "User Id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "users": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "example": "John"
          },
          "lastname": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "john@doe.com"
          },
          "password": {
            "type": "string",
            "example": "Password123*/"
          },
          "phone": {
            "type": "string",
            "example": "1234567890"
          }
        },
        "required": [
          "firstname",
          "lastname",
          "email",
          "password",
          "phone"
        ]
      }
    }
  }
}